Added the UMLS as a source of term synonyms.
This is meant to be a replacement for the Lexigram API. We wanted
more control over the vocabulary anyway but also the reviwers of
our SIGIR submission critized the use of an external black box API.
The use of the UMLS actually brought a small enhancement to the
TREC PM 2017 results and can seemingly serve as a good replacement.

+++ src/main/java/at/medunigraz/imi/bst/trec/model/Result.java
@@ -5,9 +5,9 @@ import java.io.Serializable;
 public class Result implements Serializable {
 	private String id;
 
-	private float score;
+	private double score;
 
-	public Result(String id, float score) {
+	public Result(String id, double score) {
 		this.id = id;
 		this.score = score;
 	}
@@ -16,7 +16,7 @@ public class Result implements Serializable {
 		return id;
 	}
 
-	public float getScore() {
+	public double getScore() {
 		return score;
 	}
 }
